apps/server/pyproject.toml
[tool.poetry]
name = "time-traveler"
version = "0.1.0"
description = ""
authors = ["dan1dr <dannieyo@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
elevenlabs = "^2.15.0"
python-dotenv = "^1.1.1"
ffmpeg = "^1.4"
locust = "^2.40.4"
fastapi = {version = "0.116.2", extras = ["all"]}
uvicorn = "^0.35.0"
twilio = "^9.8.0"
python-multipart = "^0.0.20"
starlette = "^0.48.0"
PyJWT = "^2.8.0"
shared-py = {path = "./shared_py", develop = true}


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.30.1"
pytest = "^8.4.2"
pytest-cov = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
